+++ Changelog.md
@@ -5,6 +5,12 @@ Enhancements:
 * Combine multiple `--pattern` arguments making them equivalent to
   `--pattern=1,2,...,n`. (Jon Rowe, #2002)
 
+Bug Fixes:
+
+* Correctly run `before(:suite)` (and friends) in the context of an example
+  group instance, thus making the expected RSpec environment available.
+  (Jon Rowe, #1986)
+
 ### 3.3.0 / 2015-06-12
 [Full Changelog](http://github.com/rspec/rspec-core/compare/v3.2.3...v3.3.0)